fronend hakee yksittäisen tietueen backendiltä
