+++ frontend/package.json
@@ -14,7 +14,10 @@
     "cors": "^2.8.5",
     "nanoid": "^5.1.5",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.0",
+    "react-xml-parser": "^1.1.8",
+    "xml2js": "^0.6.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
