+++ frontend/src/types.ts
@@ -7,6 +7,7 @@
 
 export interface Props {
   search: string,
+  server: string,
   handleSearchInputChange: (event: React.SyntheticEvent) => void,
   handleSearchEvent: (event: React.SyntheticEvent) => void,
